authorization

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -39,11 +42,14 @@ class JobController extends Controller
     }
     public function edit(Job $job)
     {
+
         return view('jobs.edit', ['job' => $job]);
     }
     public function update(Job $job)
     {
-        //authorize (On hold...)
+        //authorize
+        Gate::authorize('click-job', $job);
+
         //validate
         request()->validate([
             'title' => ['required', 'min:3'],
@@ -63,7 +69,7 @@ class JobController extends Controller
     }
     public function destroy(Job $job)
     {
-        //authorize (On hold...)
+        Gate::authorize('click-job', $job);
         $job->delete();
         return redirect('/jobs');
     }
